requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
python-pptx
